refactor: 💡 Refactor file structure
* Moved tests and files to its own directory

✅ Closes: #86du2ymne
