+++ src/infrastructure/auth/middleware/auth.middleware.ts
@@ -1,6 +1,6 @@
 import { HttpStatus, Injectable, NestMiddleware } from '@nestjs/common';
 import { Request, Response, NextFunction } from 'express';
-import { AuthService } from './auth.service';
+import { AuthService } from '../service/auth.service';
 
 @Injectable()
 export class AuthMiddleware implements NestMiddleware {
